requirements.txt
numpy==1.16.0
pandas==0.24.0
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
